fastapi_app/pyproject.toml
[tool.poetry]
name = "fastapi_app"
version = "0.1.0"
description = ""
authors = ["andrey <dred2spb79@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
aiohttp = "^3.10.10"
aiohttp-asgi = "^0.5.2"
uvicorn = "^0.32.0"
passlib = "^1.7.4"
pydantic-settings = "^2.6.1"
asyncpg = "^0.30.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
bcrypt = "^4.2.0"
python-multipart = "^0.0.17"
orjson = "^3.10.11"
redis = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
